use create_teste_sourcelist
