+++ test/variant_cast_test.cpp
@@ -35,7 +35,7 @@
 #define EPSILON 0.001
 
 ////////////////////////////////////////////////////////////////////////////////
-int TEST_NAME() {
+int TEST_NAME(int, char**) {
   UnitTest t(81);
 
   time_t now = time(nullptr);
