+++ test/variant_modulo_test.cpp
@@ -35,7 +35,7 @@
 #define EPSILON 0.0001
 
 ////////////////////////////////////////////////////////////////////////////////
-int TEST_NAME() {
+int TEST_NAME(int, char**) {
   UnitTest t(40);
 
   Variant v0(true);
