+++ test/variant_partial_test.cpp
@@ -33,7 +33,7 @@
 #include <iostream>
 
 ////////////////////////////////////////////////////////////////////////////////
-int TEST_NAME() {
+int TEST_NAME(int, char**) {
   UnitTest t(72);
 
   Variant v0(true);
@@ -42,150 +42,160 @@ int TEST_NAME() {
   Variant v3("foo");
   Variant v4(1234567890, Variant::type_date);
   Variant v5(1200, Variant::type_duration);
+  Variant v6(1234522800, Variant::type_date);  // 2009-02-13, 12.00pm UTC
+  Variant v7("2009-02-13");
 
-  Variant v00 = v0 == v0;
+  Variant v00 = v0.operator_partial(v0);
   t.is(v00.type(), Variant::type_boolean, "true == true --> boolean");
   t.is(v00.get_bool(), true, "true == true --> true");
 
-  Variant v01 = v0 == v1;
+  Variant v01 = v0.operator_partial(v1);
   t.is(v01.type(), Variant::type_boolean, "true == 42 --> boolean");
   t.is(v01.get_bool(), false, "true == 42 --> false");
 
-  Variant v02 = v0 == v2;
+  Variant v02 = v0.operator_partial(v2);
   t.is(v02.type(), Variant::type_boolean, "true == 3.14 --> boolean");
   t.is(v02.get_bool(), false, "true == 3.14 --> false");
 
-  Variant v03 = v0 == v3;
+  Variant v03 = v0.operator_partial(v3);
   t.is(v03.type(), Variant::type_boolean, "true == 'foo' --> boolean");
   t.is(v03.get_bool(), false, "true == 'foo' --> false");
 
-  Variant v04 = v0 == v4;
+  Variant v04 = v0.operator_partial(v4);
   t.is(v04.type(), Variant::type_boolean, "true == 1234567890 --> boolean");
   t.is(v04.get_bool(), false, "true == 1234567890 --> false");
 
-  Variant v05 = v0 == v5;
+  Variant v05 = v0.operator_partial(v5);
   t.is(v05.type(), Variant::type_boolean, "true == 1200 --> boolean");
   t.is(v05.get_bool(), false, "true == 1200 --> false");
 
-  Variant v10 = v1 == v0;
+  Variant v10 = v1.operator_partial(v0);
   t.is(v10.type(), Variant::type_boolean, "42 == true --> boolean");
   t.is(v10.get_bool(), false, "42 == true --> false");
 
-  Variant v11 = v1 == v1;
+  Variant v11 = v1.operator_partial(v1);
   t.is(v11.type(), Variant::type_boolean, "42 == 42 --> boolean");
   t.is(v11.get_bool(), true, "42 == 42 --> true");
 
-  Variant v12 = v1 == v2;
+  Variant v12 = v1.operator_partial(v2);
   t.is(v12.type(), Variant::type_boolean, "42 == 3.14 --> boolean");
   t.is(v12.get_bool(), false, "42 == 3.14 --> false");
 
-  Variant v13 = v1 == v3;
+  Variant v13 = v1.operator_partial(v3);
   t.is(v13.type(), Variant::type_boolean, "42 == 'foo' --> boolean");
   t.is(v13.get_bool(), false, "42 == 'foo' --> false");
 
-  Variant v14 = v1 == v4;
+  Variant v14 = v1.operator_partial(v4);
   t.is(v04.type(), Variant::type_boolean, "42 == 1234567890 --> boolean");
   t.is(v04.get_bool(), false, "42 == 1234567890 --> false");
 
-  Variant v15 = v1 == v5;
+  Variant v15 = v1.operator_partial(v5);
   t.is(v15.type(), Variant::type_boolean, "42 == 1200 --> boolean");
   t.is(v15.get_bool(), false, "42 == 1200 --> false");
 
-  Variant v20 = v2 == v0;
+  Variant v20 = v2.operator_partial(v0);
   t.is(v20.type(), Variant::type_boolean, "3.14 == true --> boolean");
   t.is(v20.get_bool(), false, "3.14 == true --> false");
 
-  Variant v21 = v2 == v1;
+  Variant v21 = v2.operator_partial(v1);
   t.is(v21.type(), Variant::type_boolean, "3.14 == 42 --> boolean");
   t.is(v21.get_bool(), false, "3.14 == 42 --> false");
 
-  Variant v22 = v2 == v2;
+  Variant v22 = v2.operator_partial(v2);
   t.is(v22.type(), Variant::type_boolean, "3.14 == 3.14 --> boolean");
   t.is(v22.get_bool(), true, "3.14 == 3.14 --> true");
 
-  Variant v23 = v2 == v3;
+  Variant v23 = v2.operator_partial(v3);
   t.is(v23.type(), Variant::type_boolean, "3.14 == 'foo' --> boolean");
   t.is(v23.get_bool(), false, "3.14 == 'foo' --> false");
 
-  Variant v24 = v2 == v4;
+  Variant v24 = v2.operator_partial(v4);
   t.is(v24.type(), Variant::type_boolean, "3.14 == 1234567890 --> boolean");
   t.is(v24.get_bool(), false, "3.14 == 1234567890 --> false");
 
-  Variant v25 = v2 == v5;
+  Variant v25 = v2.operator_partial(v5);
   t.is(v25.type(), Variant::type_boolean, "3.14 == 1200 --> boolean");
   t.is(v25.get_bool(), false, "3.14 == 1200 --> false");
 
-  Variant v30 = v3 == v0;
+  Variant v30 = v3.operator_partial(v0);
   t.is(v30.type(), Variant::type_boolean, "'foo' == true --> boolean");
   t.is(v30.get_bool(), false, "'foo' == true --> false");
 
-  Variant v31 = v3 == v1;
+  Variant v31 = v3.operator_partial(v1);
   t.is(v31.type(), Variant::type_boolean, "'foo' == 42 --> boolean");
   t.is(v31.get_bool(), false, "'foo' == 42 --> false");
 
-  Variant v32 = v3 == v2;
+  Variant v32 = v3.operator_partial(v2);
   t.is(v32.type(), Variant::type_boolean, "'foo' == 3.14 --> boolean");
   t.is(v32.get_bool(), false, "'foo' == 3.14 --> false");
 
-  Variant v33 = v3 == v3;
+  Variant v33 = v3.operator_partial(v3);
   t.is(v33.type(), Variant::type_boolean, "'foo' == 'foo' --> boolean");
   t.is(v33.get_bool(), true, "'foo' == 'foo' --> true");
 
-  Variant v34 = v3 == v4;
+  Variant v34 = v3.operator_partial(v4);
   t.is(v34.type(), Variant::type_boolean, "'foo' == 1234567890 --> boolean");
   t.is(v34.get_bool(), false, "'foo' == 1234567890 --> false");
 
-  Variant v35 = v3 == v5;
+  Variant v35 = v3.operator_partial(v5);
   t.is(v35.type(), Variant::type_boolean, "'foo' == 1200 --> boolean");
   t.is(v35.get_bool(), false, "'foo' == 1200 --> false");
 
-  Variant v40 = v4 == v0;
+  Variant v40 = v4.operator_partial(v0);
   t.is(v40.type(), Variant::type_boolean, "1234567890 == true --> boolean");
   t.is(v40.get_bool(), false, "1234567890 == true --> false");
 
-  Variant v41 = v4 == v1;
+  Variant v41 = v4.operator_partial(v1);
   t.is(v41.type(), Variant::type_boolean, "1234567890 == 42 --> boolean");
   t.is(v41.get_bool(), false, "1234567890 == 42 --> false");
 
-  Variant v42 = v4 == v2;
+  Variant v42 = v4.operator_partial(v2);
   t.is(v42.type(), Variant::type_boolean, "1234567890 == 3.14 --> boolean");
   t.is(v42.get_bool(), false, "1234567890 == 3.14 --> false");
 
-  Variant v43 = v4 == v3;
+  Variant v43 = v4.operator_partial(v3);
   t.is(v43.type(), Variant::type_boolean, "1234567890 == 'foo' --> boolean");
   t.is(v43.get_bool(), false, "1234567890 == 'foo' --> false");
 
-  Variant v44 = v4 == v4;
+  Variant v44 = v4.operator_partial(v4);
   t.is(v44.type(), Variant::type_boolean, "1234567890 == 1234567890 --> boolean");
   t.is(v44.get_bool(), true, "1234567890 == 1234567890 --> true");
 
-  Variant v45 = v4 == v5;
+  Variant v45 = v4.operator_partial(v5);
   t.is(v45.type(), Variant::type_boolean, "1234567890 == 1200 --> boolean");
   t.is(v45.get_bool(), false, "1234567890 == 1200 --> false");
 
-  Variant v50 = v5 == v0;
+  Variant v50 = v5.operator_partial(v0);
   t.is(v50.type(), Variant::type_boolean, "1200 == true --> boolean");
   t.is(v50.get_bool(), false, "1200 == true --> false");
 
-  Variant v51 = v5 == v1;
+  Variant v51 = v5.operator_partial(v1);
   t.is(v51.type(), Variant::type_boolean, "1200 == 42 --> boolean");
   t.is(v51.get_bool(), false, "1200 == 42 --> false");
 
-  Variant v52 = v5 == v2;
+  Variant v52 = v5.operator_partial(v2);
   t.is(v52.type(), Variant::type_boolean, "1200 == 3.14 --> boolean");
   t.is(v52.get_bool(), false, "1200 == 3.14 --> false");
 
-  Variant v53 = v5 == v3;
+  Variant v53 = v5.operator_partial(v3);
   t.is(v53.type(), Variant::type_boolean, "1200 == 'foo' --> boolean");
   t.is(v53.get_bool(), false, "1200 == 'foo' --> false");
 
-  Variant v54 = v5 == v4;
+  Variant v54 = v5.operator_partial(v4);
   t.is(v04.type(), Variant::type_boolean, "1200 == 1234567890 --> boolean");
   t.is(v04.get_bool(), false, "1200 == 1234567890 --> false");
 
-  Variant v55 = v5 == v5;
+  Variant v55 = v5.operator_partial(v5);
   t.is(v55.type(), Variant::type_boolean, "1200 == 1200 --> boolean");
   t.is(v55.get_bool(), true, "1200 == 1200 --> true");
+
+  Variant v56 = v6.operator_partial(v7);
+  t.is(v56.type(), Variant::type_boolean, "1234522800 == '2009-02-13' --> boolean");
+  t.is(v56.get_bool(), true, "1234522800 == '2009-02-13' --> true");
+
+  Variant v57 = v7.operator_partial(v6);
+  t.is(v57.type(), Variant::type_boolean, "'2009-02-13' == 1234522800 --> boolean");
+  t.is(v57.get_bool(), true, "'2009-02-13' == 1234522800 --> true");
 
   return 0;
 }
